+++ internal/models/withdraw.go
@@ -8,8 +8,8 @@ type WithdrawalsJSON struct {
 	UserID      uint64    `gorm:"not null" json:"-"`
 	OrderNumber string    `gorm:"unique;not null" json:"order"`
 	Status      string    `gorm:"default:'NEW'" json:"-"`
-	Accrual     float64   `gorm:"default:null" json:"-"`
-	Sum         float64   `gorm:"default:null" json:"sum"`
+	Accrual     float32   `gorm:"default:null" json:"-"`
+	Sum         float32   `gorm:"default:null" json:"sum"`
 	UploadedAt  time.Time `gorm:"autoCreateTime" json:"-"`
 	UpdatedAt   time.Time `gorm:"autoUpdateTime" json:"processed_at"`
 }
